That was awful
I tried to fix it quickly, which changed to long long with code duplication which is now impossible to navigate. I know how to fix it - by introducing an object which would expose information needed for a given recipe. But this requires a lot of refactor, and I got so tired with this work, this addon, that I do not know whether I will change it. Maybe... at some point. Bit by bit. I am happy it is over. Now I need to clear it up. And maybe change finishing reagents so they could be applied also at earlier stages to reach tiers not just at the end.
I am sorry if anyone ever reads this. I hope nobody will ;)